Format numbers

The function fmt_new creates a function to format numbers using a given locale and spec. The function fmt formats numbers, and is a convenience function for fmt_new(...)(x).

fmt_new(spec = NULL, locale = NULL, si = NULL)

fmt(x, spec = NULL, locale = NULL, si = NULL)

Arguments

spec

A fmt_spec object, or a string coerced to a fmt_spec object using as_fmt_spec, or a list of arguments passed to fmt_spec.

locale

A fmt_locale object.

si

NULL or a numeric vector with length one. The SI-prefix to use x. Unlike the "s" format type, this applies the same prefix to all values in x, rather than determining the SI-prefix on a value by value basis. Additionally, a non-NULL si implies a "f" format type, in which precision represents the number of digits past the decimal point.

x

A numeric or integer vector

Value

fmt_new returns a function with a single argument. fmt returns a function of the same length as x of formatted numbers.

Details

The supported SI-prefixes are:

"y" - yocto, \(10^{-24}\)

"z" - zepto, \(10^{-21}\)

"a" - atto, \(10^{-18}\)

"f" - femto, \(10^{-15}\)

"p" - pico, \(10^{-12}\)

"n" - nano, \(10^{-9}\)

"µ" - micro, \(10^{-6}\)

"m" - milli, \(10^{-3}\)

" " (none) - \(10^0\)

"k" - kilo, \(10^3\)

"M" - mega, \(10^{6}\)

"G" - giga, \(10^{9}\)

"T" - tera, \(10^{12}\)

"P" - peta, \(10^{15}\)

"E" - exa, \(10^{18}\)

"Z" - zetta, \(10^{21}\)

"Y" - yotta, \(10^{24}\)

Examples

fmt(c(0.00042, 0.0042), spec = ",.0", si = 1e-6)

#> [1] "420μ"   "4,200μ"