Format Specification

Format Specification

as_fmt_spec(x = character())

fmt_spec(fill = " ", align = c(">", "<", "^", "="), sign = c("-", "+",
  "(", " "), symbol = NULL, zero = FALSE, width = NULL, comma = FALSE,
  precision = NULL, type = NULL)

Arguments

x

A string representation of the spec. See Details.

fill

A string. See Details.

align

A string. See Details.

sign

A string. See Details.

symbol

A string. See Details.

zero

A logical vector of length one. See Details.

width

An integer vector of length one. See Details.

comma

An logical vector of length one. See Details.

precision

An integer vector of length one. See Details.

type

A character vector of length one. See Details.

Value

An object of class "fmt_spec". This is a list with elements:

fill

Character vector of length one. See Details.

align

Character vector of length one. See Details.

sign

Character vector of length one. See Details.

symbol

Character vector of length one. See Details.

width

Integer vector of length one. See Details.

comma

Logical vector of length one. See Details.

precision

Integer vector of length one.

type

Character vector of length one, or NULL if the default type

Details

as_fmt_spec() returns a fmt_spec object for the given string specifier. The returned function takes a number as the only argument, and returns a string representing the formatted number. The general form of a specifier is: [[fill]align][sign][symbol][0][width][,][.precision][type]

The fill can be any character. The presence of a fill character is signaled by the *align* character following it, which must be one of the following:

>: Forces the field to be right-aligned within the available space.

<: Forces the field to be left-aligned within the available space.

^: Forces the field to be centered within the available space.

=: like >, but with any sign and symbol to the left of any padding.

The sign can be:

"-} : nothing for positive and a minus sign for negative. (Default behavior.)} \item{\code{"+"} : a plus sign for positive and a minus sign for negative.} \code{{"("} : nothing for positive and parentheses for negative.} \code{{" "} (space) : a space for positive and a minus sign for negative.} }

The \code{symbol} can be: \itemize{ \item{\code{"$"}: apply currency symbols per the locale definition.} \item{\code{"#": for binary, octal, or hexadecimal notation, prefix by "0b", "0o", or "0x", respectively.

The zero (0) option enables zero-padding; this implicitly sets fill to "0" and align to "=". The width defines the minimum field width; if not specified, then the width will be determined by the content. The comma (,) option enables the use of a group separator, such as a comma for thousands.

Depending on the type, the precision either indicates the number of digits that follow the decimal point ("f", "%"), or the number of significant digits (NULL, "e"`, `"g"`, `"r"`, `"s"`, `"p"`). If the precision is not specified, it defaults to 6 for all types except NULL, which defaults to 12. Precision is ignored for integer formats (types "b", "o", "d", "x", "X" and "c"). See precision_fixed and precision_round for help picking an appropriate precision.

The available type values are:

"e" - exponent notation.

"f" - fixed point notation.

"g" - either decimal or exponent notation, rounded to significant digits.

"r" - decimal notation, rounded to significant digits.

"s" - decimal notation with an [SI prefix](#locale_formatPrefix), rounded to significant digits.

"%" - multiply by 100, and then decimal notation with a percent sign.

"p" - multiply by 100, round to significant digits, and then decimal notation with a percent sign.

"b" - binary notation, rounded to integer.

"o" - octal notation, rounded to integer.

"d" - decimal notation, rounded to integer.

"x" - hexadecimal notation, using lower-case letters, rounded to integer.

"X" - hexadecimal notation, using upper-case letters, rounded to integer.

"c" - converts the integer to the corresponding unicode character before printing.

NULL - like "g", but trim insignificant trailing zeros.

The type n is also supported as shorthand for ,g.

For the "g", "n" and NULL (none) types, decimal notation is used if the resulting string would have precision or fewer digits; otherwise, an exponent notation is used.